requirements.txt
requests
beautifulsoup4
lxml
PyYAML
playwright
flask
PyPDF2>=3.0.0
python-docx>=0.8.11
schedule>=1.2.0
